fix: improve companion lifecycle management and test coverage
- Refactor companion IDM addition tracking with SetAddedToIdm() API
- Remove automatic timeout cleanup for companions not added to IDM
- Add device ID validation and message size buffer checks
- Update available device subscription to monitor companion status changes
- Add templateId to EndAddCompanionOutput for proper tracking
- Remove CFI-problematic test files (inbound/outbound_request_test)
- Re-enable message_router_test and update test initialization
- Standardize LOG_TAG to "CDA_SA" across multiple modules
- Fix clang-format formatting issues

+++ services/service_entry/src/subscription/subscription_util.cpp
@@ -45,8 +45,8 @@ IpcTemplateStatus ConvertToIpcTemplateStatus(const CompanionStatus &companionSta
 {
     IpcTemplateStatus ipcStatus {};
     ipcStatus.templateId = companionStatus.templateId;
-    ipcStatus.isConfirmed =
-        manageSubscribeTime.has_value() && (companionStatus.lastCheckTime >= manageSubscribeTime.value());
+    ipcStatus.isConfirmed = manageSubscribeTime.has_value() &&
+        (companionStatus.lastCheckTime >= static_cast<uint64_t>(manageSubscribeTime.value()));
     ipcStatus.isValid = companionStatus.isValid;
     ipcStatus.localUserId = companionStatus.hostUserId;
     ipcStatus.addedTime = companionStatus.addedTime;
